+++ src/impls/mod.rs
@@ -1,6 +1,5 @@
 #[cfg(target_os = "android")]
 mod android;
-
 #[cfg(target_os = "android")]
 use android::AndroidFsImpl;